Macro refactoring
